🔥 ♻️  refactor code( useless files removed)
