--- routes/offer-listings/index.jsx
@@ -1,13 +1,0 @@
-import React from 'react'
-import { Layout } from "../../components"
-import OfferListings from "./OfferListings"
-
-const Index = () => {
-    return (
-        <Layout>
-            <OfferListings />
-        </Layout>
-    )
-}
-
-export default Index
