+++ styles/_components/_layout.scss
@@ -4,7 +4,14 @@
   padding: 50px 8%;
   header {
     margin-bottom: 80px;
-    a {
+
+    button {
+      border: none;
+      background-color: transparent;
+    }
+
+    a,
+    button {
       font-weight: 500;
       font-size: 1rem;
       color: $white;
